+++ entity/product-order.go
@@ -1,7 +1,7 @@
 package entity
 
 type ProductOrder struct {
-	DiscountAmount float32 `json:"discount_amount"`
+	DiscountAmount float64 `json:"discount_amount"`
 	DiscountType   string  `json:"discount_type"`
 	Ean            string  `json:"ean"`
 	Isbn           string  `json:"isbn"`
@@ -11,14 +11,14 @@ type ProductOrder struct {
 	Mpn            string  `json:"mpn"`
 	Name           string  `json:"name"`
 	OrderId        int64   `json:"order_id"`
-	Price          float32 `json:"price"`
+	Price          float64 `json:"price"`
 	ProductId      int64   `json:"product_id"`
 	ProductUid     string  `json:"product_uid"`
-	Quantity       float32 `json:"quantity"`
-	Reward         float32 `json:"reward"`
+	Quantity       float64 `json:"quantity"`
+	Reward         float64 `json:"reward"`
 	Sku            string  `json:"sku"`
-	Tax            float32 `json:"tax"`
-	Total          float32 `json:"total"`
+	Tax            float64 `json:"tax"`
+	Total          float64 `json:"total"`
 	Upc            string  `json:"upc"`
-	Weight         float32 `json:"weight"`
+	Weight         float64 `json:"weight"`
 }
