+++ entity/product-special.go
@@ -9,7 +9,7 @@ import (
 type ProductSpecial struct {
 	ProductUid string    `json:"product_uid" validate:"required"`
 	GroupId    int64     `json:"group_id" validate:"required,number,gt=0"`
-	Price      float32   `json:"price" validate:"required,number,gt=0"`
+	Price      float64   `json:"price" validate:"required,number,gt=0"`
 	Priority   int       `json:"priority" validate:"omitempty,number"`
 	DateStart  time.Time `json:"date_start" validate:"omitempty"`
 	DateEnd    time.Time `json:"date_end" validate:"omitempty"`
